# More than double adler32 performance with altivec