@mtl1979 2022-01-28
Seems like qemu-ppc is not supporting VMX at all, so the coverage is 0%... I might need to dig my old Ubuntu/PPC and check if I can get some adler32 benchmark comparisons done...