@KungFuJesus 2022-01-28
Please do.  I have half a mind to try to unroll this a bit further as everything I've read seems to indicate the G5 has a pretty deeply staged pipeline and to maximize the throughput I'm going to need to pop some more data dependency bubbles.  But it could ultimately by that the dependent loads from memory are going to be the limiting factor.  At the very least I can unroll the the weird cache line size of the G5 (128 bytes).