@KungFuJesus 2022-01-29
```
Comparing adler32/vmx (from ./benchmark_zlib_old) to adler32/vmx (from ./benchmark_zlib)
Benchmark                                                      Time             CPU      Time Old      Time New       CPU Old       CPU New
-------------------------------------------------------------------------------------------------------------------------------------------
[adler32/vmx vs. adler32/vmx]/2048_mean                     -0.6817         -0.6817          1159           369          1159           369
[adler32/vmx vs. adler32/vmx]/2048_median                   -0.6816         -0.6816          1159           369          1159           369
[adler32/vmx vs. adler32/vmx]/2048_stddev                   -0.1940         +0.7143             0             0             0             0
[adler32/vmx vs. adler32/vmx]/4096_mean                     -0.6927         -0.6926          2286           703          2286           703
[adler32/vmx vs. adler32/vmx]/4096_median                   -0.6926         -0.6926          2286           703          2286           703
[adler32/vmx vs. adler32/vmx]/4096_stddev                   -0.7952         -0.7431             0             0             0             0
[adler32/vmx vs. adler32/vmx]/32768_mean                    -0.6953         -0.6953         18157          5532         18156          5532
[adler32/vmx vs. adler32/vmx]/32768_median                  -0.6953         -0.6953         18156          5532         18156          5532
[adler32/vmx vs. adler32/vmx]/32768_stddev                  -0.7204         -0.4717             2             1             1             0
[adler32/vmx vs. adler32/vmx]/262144_mean                   -0.6961         -0.6961        145631         44257        145622         44256
[adler32/vmx vs. adler32/vmx]/262144_median                 -0.6961         -0.6961        145629         44257        145620         44256
[adler32/vmx vs. adler32/vmx]/262144_stddev                 -0.8199         -0.7243            13             2             8             2
[adler32/vmx vs. adler32/vmx]/2097152_mean                  -0.5188         -0.5188       1290244        620873       1290164        620858
[adler32/vmx vs. adler32/vmx]/2097152_median                -0.5187         -0.5187       1290254        620964       1290157        620952
[adler32/vmx vs. adler32/vmx]/2097152_stddev                +0.2937         +0.6945           209           271           158           268
[adler32/vmx vs. adler32/vmx]/4194304_mean                  -0.5028         -0.5028       2588475       1286875       2588345       1286849
[adler32/vmx vs. adler32/vmx]/4194304_median                -0.5028         -0.5028       2588463       1286932       2588365       1286905
[adler32/vmx vs. adler32/vmx]/4194304_stddev                -0.3738         -0.1664           538           337           405           337
```