@Dead2 2022-01-30
Ohh, I had partly forgotten about that. Lets just ignore that test for now.

We moved the last of the servers out of my jobs server room on Friday.
That specific server (Snipp) is not going to come back up (Dual Xeon Westmere is more power hungry than I want to be hosting from home on my own money, especially after power prices went up 5-10x the past 4 months)..
I do have a haswell machine that could be used perhaps, but that will probably take some time.

Btw, I just spent ~6 hours setting up an armv6l machine for our testing purposes, that and the aarch64 one might possibly be added to CI at some point. They are pretty slow though, so probably not good as complete replacement for the ones currently running CI via qemu. They are based on RPI 1B and RPI 3, would have been nice to have two RPI 4 (or better) to use instead, but we will have to make do.