@KungFuJesus 2022-01-31
In theory, anyway, CI should be testing it.  For benchmarking the gains - it sounds like @mtl1979 has a set of PPC hardware that he can test with if you need a second data point.