@KungFuJesus 2022-01-31
> When I wrote the original vectorized version, I was still learning the instruction set and the other intricacies of CPU internals. My highest priority was keeping the code as readable as possible without deviating too much from the original code. That was to attract more experienced developers to improve the code.

I get it - my first cut of Altivec code does some suboptimal things around unaligned loads when I first learned.  I'm suggesting you can benchmark this PR with your hardware so that we can get a second data point.