@mtl1979 2022-01-31
> I get it - my first cut of Altivec code does some suboptimal things around unaligned loads when I first learned. I'm suggesting you can benchmark this PR with your hardware so that we can get a second data point.

I just need to make some room so I can connect everything again... Doesn't really help that I have two cats here jumping around, not forgetting all the other projects I'm working on currently.