@mtl1979 2022-01-31
Building benchmark_zlib requires more recent cmake version than my 32-bit Ubuntu (16.10) has... There is no later release for 32-bit PowerPC.