@mtl1979 2022-02-01
```
before:

--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
adler32/vmx/2048          6720 ns         6717 ns       105787
adler32/vmx/4096         13358 ns        13345 ns        52555
adler32/vmx/32768       105754 ns       105705 ns         6590
adler32/vmx/262144      848151 ns       846999 ns          821
adler32/vmx/2097152    6995895 ns      6990538 ns          100
adler32/vmx/4194304   14016579 ns     13949114 ns           50

after:

--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
adler32/vmx/2048          2963 ns         2959 ns       236459
adler32/vmx/4096          5853 ns         5839 ns       119032
adler32/vmx/32768        45709 ns        45612 ns        15529
adler32/vmx/262144      363820 ns       363571 ns         1930
adler32/vmx/2097152    3110848 ns      3108457 ns          227
adler32/vmx/4194304    6199476 ns      6194799 ns          112
```