@mscastanho 2022-02-01
> There is no later release for 32-bit PowerPC.

@mtl1979 Debian is one of the only distros that still support 32-bit PowerPC, but only in the unstable branch (Debian sid) [Link](http://deb.debian.org/debian-ports/dists/sid/).